+++ packages/skills/carpark_client/skill.yaml
@@ -7,15 +7,15 @@ behaviours:
   - behaviour:
       class_name: MySearchBehaviour
       args: {}
-  - behaviour:
-      class_name: MyTransactionBehaviour
-      args: {}
 handlers:
   - handler:
       class_name: FIPAHandler
       args: {}
   - handler:
       class_name: OEFHandler
+      args: {}
+  - handler:
+      class_name: MyTransactionHandler
       args: {}
 tasks: []
 shared_classes:
